requirements.txt
#### REQ ####
# mysql-connector-python
python-dotenv
datetime
pandas
pmaw
pymysql
cherrypicker